fix: resolve medium and low priority bugs from audit report

+++ database/migrations/2026_01_04_040400_create_variant_sub_processes_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('product_variant_id')->constrained('product_variants')->onDelete('cascade');
             $table->foreignId('sub_process_id')->constrained('sub_processes')->onDelete('cascade');
             $table->integer('sequence_order')->default(1);
-            $table->decimal('weight', 5, 2)->default(1.00);
+            $table->integer('weight')->default(1);
             $table->timestamps();
             
             $table->unique(['product_variant_id', 'sub_process_id']);
